pyproject.toml
[project]
name = "laocai-qqbot"
version = "2.0.0"
description = "基于FastAPI的智能QQ群机器人，支持AI聊天和掷骰功能"
readme = "README.md"
requires-python = ">=3.12, <3.13"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "cryptography>=45.0.2",
    "fastapi[standard]>=0.115.0",
    "openai>=1.81.0",
    "requests>=2.32.3",
    "pyyaml>=6.0.0",
    "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

[project.scripts]
laocai-bot = "main_new:main"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
